Fixed #24564 - Moved AbstractBaseUser and BaseUserManager so they can be used without auth in INSTALLED_APPS
